src/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Naked==0.1.31
pycryptodome==3.11.0
pycryptodomex==3.11.0
PyYAML==6.0
requests==2.26.0
shellescape==3.8.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.5.0